pyproject.toml
[project]
name = "leadops"
version = "0.1.0"
description = "Local-first lead operations with an Airtable mirror."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "requests>=2.31.0",
  "openpyxl>=3.1.2",
]

[project.scripts]
crm = "crm.cli:app"

[build-system]
requires = ["setuptools>=68.2.2", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["crm*"]
exclude = ["data*", "exports*", "schema*", "workspaces*", "templates*", "docs*", "tests*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".pixi", ".venv", "data", "exports", "workspaces"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
